+++ packages/lib-sourcify/src/SourcifyChain/SourcifyChainTypes.ts
@@ -12,7 +12,7 @@ export type SourcifyChainExtension = {
   sourcifyName: string; // Keep it required to not forget name in sourcify-chains.json
   supported: boolean;
   etherscanApi?: {
-    apiURL: string;
+    supported: boolean;
     apiKeyEnvName?: string;
   };
   fetchContractCreationTxUsing?: FetchContractCreationTxMethods;
@@ -90,6 +90,7 @@ export interface FetchContractCreationTxMethods {
     url: string;
     runtime: string;
   };
+  veChainApi?: boolean;
 }
 
 export type Chain = {