+++ services/server/src/server/apiv2/verification/verification.handlers.ts
@@ -4,11 +4,14 @@ import type {
   CompilationTarget,
   Metadata,
 } from "@ethereum-sourcify/lib-sourcify";
+import { splitFullyQualifiedName } from "@ethereum-sourcify/lib-sourcify";
 import { TypedResponse } from "../../types";
 import logger from "../../../common/logger";
 import { Request } from "express";
 import { Services } from "../../services/services";
 import { StatusCodes } from "http-status-codes";
+import { fetchFromEtherscan } from "../../services/utils/etherscan-util";
+import { ChainRepository } from "../../../sourcify-chain-repository";
 
 interface VerifyFromJsonInputRequest extends Request {
   params: {
@@ -23,18 +26,17 @@ interface VerifyFromJsonInputRequest extends Request {
   };
 }
 
-type VerifyFromJsonInputResponse = TypedResponse<{
+type VerifyResponse = TypedResponse<{
   verificationId: string;
 }>;
 
 export async function verifyFromJsonInputEndpoint(
   req: VerifyFromJsonInputRequest,
-  res: VerifyFromJsonInputResponse,
+  res: VerifyResponse,
 ) {
   logger.debug("verifyFromJsonInputEndpoint", {
     chainId: req.params.chainId,
     address: req.params.address,
-    stdJsonInput: req.body.stdJsonInput,
     compilerVersion: req.body.compilerVersion,
     contractIdentifier: req.body.contractIdentifier,
     creationTransactionHash: req.body.creationTransactionHash,
@@ -42,9 +44,9 @@ export async function verifyFromJsonInputEndpoint(
 
   // The contract path can include a colon itself. Therefore,
   // we need to take the last element as the contract name.
-  const splitIdentifier = req.body.contractIdentifier.split(":");
-  const contractName = splitIdentifier[splitIdentifier.length - 1];
-  const contractPath = splitIdentifier.slice(0, -1).join(":");
+  const { contractName, contractPath } = splitFullyQualifiedName(
+    req.body.contractIdentifier,
+  );
   const compilationTarget: CompilationTarget = {
     name: contractName,
     path: contractPath,
@@ -77,13 +79,9 @@ interface VerifyFromMetadataRequest extends Request {
   };
 }
 
-type VerifyFromMetadataResponse = TypedResponse<{
-  verificationId: string;
-}>;
-
 export async function verifyFromMetadataEndpoint(
   req: VerifyFromMetadataRequest,
-  res: VerifyFromMetadataResponse,
+  res: VerifyResponse,
 ) {
   logger.debug("verifyFromMetadataEndpoint", {
     chainId: req.params.chainId,
@@ -102,6 +100,48 @@ export async function verifyFromMetadataEndpoint(
       req.body.metadata,
       req.body.sources,
       req.body.creationTransactionHash,
+    );
+
+  res.status(StatusCodes.ACCEPTED).json({ verificationId });
+}
+
+interface VerifyFromEtherscanRequest extends Request {
+  params: {
+    chainId: string;
+    address: string;
+  };
+  body: {
+    apiKey?: string;
+  };
+}
+
+export async function verifyFromEtherscanEndpoint(
+  req: VerifyFromEtherscanRequest,
+  res: VerifyResponse,
+) {
+  logger.debug("verifyFromEtherscanEndpoint", {
+    chainId: req.params.chainId,
+    address: req.params.address,
+  });
+
+  const services = req.app.get("services") as Services;
+  const chainRepository = req.app.get("chainRepository") as ChainRepository;
+
+  // Fetch here to give early feedback to the user.
+  // Then, process in worker.
+  const etherscanResult = await fetchFromEtherscan(
+    chainRepository.supportedChainMap[req.params.chainId],
+    req.params.address,
+    req.body?.apiKey,
+    true,
+  );
+
+  const verificationId =
+    await services.verification.verifyFromEtherscanViaWorker(
+      req.baseUrl + req.path,
+      req.params.chainId,
+      req.params.address,
+      etherscanResult,
     );
 
   res.status(StatusCodes.ACCEPTED).json({ verificationId });