+++ services/server/test/integration/apiv2/verification/verify.json.spec.ts
@@ -8,8 +8,12 @@ import { LocalChainFixture } from "../../../helpers/LocalChainFixture";
 import { ServerFixture } from "../../../helpers/ServerFixture";
 import path from "path";
 import fs from "fs";
-import { assertContractSaved } from "../../../helpers/assertions";
+import { assertJobVerification } from "../../../helpers/assertions";
 import sinon from "sinon";
+import {
+  testAlreadyBeingVerified,
+  testAlreadyVerified,
+} from "../../../helpers/common-tests";
 
 chai.use(chaiHttp);
 
@@ -21,6 +25,21 @@ describe("POST /v2/verify/:chainId/:address", function () {
 
   afterEach(async () => {
     sandbox.restore();
+  });
+
+  it("should return an invalid_json error if the body JSON is invalid", async () => {
+    const verifyRes = await chai
+      .request(serverFixture.server.app)
+      .post(
+        `/v2/verify/${chainFixture.chainId}/${chainFixture.defaultContractAddress}`,
+      )
+      .set("Content-Type", "application/json")
+      .send("{ invalid-json }");
+
+    chai.expect(verifyRes.status).to.equal(400);
+    chai.expect(verifyRes.body.customCode).to.equal("invalid_json");
+    chai.expect(verifyRes.body).to.have.property("errorId");
+    chai.expect(verifyRes.body).to.have.property("message");
   });
 
   it("should verify a contract with Solidity standard input JSON", async () => {
@@ -41,68 +60,13 @@ describe("POST /v2/verify/:chainId/:address", function () {
         creationTransactionHash: chainFixture.defaultContractCreatorTx,
       });
 
-    chai.expect(verifyRes.status).to.equal(202);
-    chai.expect(verifyRes.body).to.have.property("verificationId");
-    chai
-      .expect(verifyRes.body.verificationId)
-      .to.match(
-        /^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$/,
-      );
-
-    const jobRes = await chai
-      .request(serverFixture.server.app)
-      .get(`/v2/verify/${verifyRes.body.verificationId}`);
-
-    chai.expect(jobRes.status).to.equal(200);
-    chai.expect(jobRes.body).to.deep.include({
-      isJobCompleted: false,
-      verificationId: verifyRes.body.verificationId,
-      contract: {
-        match: null,
-        creationMatch: null,
-        runtimeMatch: null,
-        chainId: chainFixture.chainId,
-        address: chainFixture.defaultContractAddress,
-      },
-    });
-    chai.expect(jobRes.body.error).to.be.undefined;
-
-    await resolveWorkers();
-
-    const jobRes2 = await chai
-      .request(serverFixture.server.app)
-      .get(`/v2/verify/${verifyRes.body.verificationId}`);
-
-    const verifiedContract = {
-      match: "exact_match",
-      creationMatch: "exact_match",
-      runtimeMatch: "exact_match",
-      chainId: chainFixture.chainId,
-      address: chainFixture.defaultContractAddress,
-    };
-
-    chai.expect(jobRes2.status).to.equal(200);
-    chai.expect(jobRes2.body).to.include({
-      isJobCompleted: true,
-      verificationId: verifyRes.body.verificationId,
-    });
-    chai.expect(jobRes2.body.error).to.be.undefined;
-    chai.expect(jobRes2.body.contract).to.include(verifiedContract);
-
-    const contractRes = await chai
-      .request(serverFixture.server.app)
-      .get(
-        `/v2/contract/${chainFixture.chainId}/${chainFixture.defaultContractAddress}`,
-      );
-
-    chai.expect(contractRes.status).to.equal(200);
-    chai.expect(contractRes.body).to.include(verifiedContract);
-
-    await assertContractSaved(
-      serverFixture.sourcifyDatabase,
-      chainFixture.defaultContractAddress,
+    await assertJobVerification(
+      serverFixture,
+      verifyRes,
+      resolveWorkers,
       chainFixture.chainId,
-      "perfect",
+      chainFixture.defaultContractAddress,
+      "exact_match",
     );
   });
 
@@ -137,7 +101,6 @@ describe("POST /v2/verify/:chainId/:address", function () {
       .request(serverFixture.server.app)
       .post(`/v2/verify/${chainFixture.chainId}/${contractAddress}`)
       .send({
-        language: "Vyper",
         stdJsonInput: {
           language: "Vyper",
           sources: {
@@ -157,56 +120,13 @@ describe("POST /v2/verify/:chainId/:address", function () {
         creationTransactionHash: txHash,
       });
 
-    chai.expect(verifyRes.status).to.equal(202);
-    chai.expect(verifyRes.body).to.have.property("verificationId");
-    chai
-      .expect(verifyRes.body.verificationId)
-      .to.match(
-        /^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$/,
-      );
-
-    const jobRes = await chai
-      .request(serverFixture.server.app)
-      .get(`/v2/verify/${verifyRes.body.verificationId}`);
-
-    chai.expect(jobRes.status).to.equal(200);
-    chai.expect(jobRes.body).to.deep.include({
-      isJobCompleted: false,
-      verificationId: verifyRes.body.verificationId,
-      contract: {
-        match: null,
-        creationMatch: null,
-        runtimeMatch: null,
-        chainId: chainFixture.chainId,
-        address: contractAddress,
-      },
-    });
-    chai.expect(jobRes.body.error).to.be.undefined;
-
-    await resolveWorkers();
-
-    const jobRes2 = await chai
-      .request(serverFixture.server.app)
-      .get(`/v2/verify/${verifyRes.body.verificationId}`);
-
-    chai.expect(jobRes2.status).to.equal(200);
-    chai.expect(jobRes2.body).to.include({
-      isJobCompleted: true,
-      verificationId: verifyRes.body.verificationId,
-    });
-    chai.expect(jobRes2.body.contract).to.include({
-      match: "match",
-      creationMatch: "match",
-      runtimeMatch: "match",
-      chainId: chainFixture.chainId,
-      address: contractAddress,
-    });
-
-    await assertContractSaved(
-      serverFixture.sourcifyDatabase,
-      contractAddress,
+    await assertJobVerification(
+      serverFixture,
+      verifyRes,
+      resolveWorkers,
       chainFixture.chainId,
-      "partial",
+      contractAddress,
+      "match",
     );
   });
 
@@ -227,29 +147,17 @@ describe("POST /v2/verify/:chainId/:address", function () {
         )[0].join(":"),
       });
 
-    chai.expect(verifyRes.status).to.equal(202);
-
-    await resolveWorkers();
-
-    const jobRes = await chai
-      .request(serverFixture.server.app)
-      .get(`/v2/verify/${verifyRes.body.verificationId}`);
-
-    chai.expect(jobRes.status).to.equal(200);
-    chai.expect(jobRes.body).to.include({
-      isJobCompleted: true,
-    });
-    chai.expect(jobRes.body.error).to.be.undefined;
-    chai.expect(jobRes.body.contract).to.include({
-      match: "exact_match",
-      creationMatch: "exact_match",
-      runtimeMatch: "exact_match",
-      chainId: chainFixture.chainId,
-      address: chainFixture.defaultContractAddress,
-    });
+    await assertJobVerification(
+      serverFixture,
+      verifyRes,
+      resolveWorkers,
+      chainFixture.chainId,
+      chainFixture.defaultContractAddress,
+      "exact_match",
+    );
   });
 
-  it("should store a job error if the compiler returns an error", async () => {
+  it("should store a job error if the Solidity compiler returns an error", async () => {
     const { resolveWorkers } = makeWorkersWait();
 
     const sourcePath = Object.keys(
@@ -292,6 +200,11 @@ describe("POST /v2/verify/:chainId/:address", function () {
     });
     chai.expect(jobRes.body.error).to.exist;
     chai.expect(jobRes.body.error.customCode).to.equal("compiler_error");
+    chai
+      .expect(jobRes.body.error.errorData.compilerErrors[0].formattedMessage)
+      .to.equal(
+        "ParserError: Expected ';' but got '{'\n --> project:/contracts/Storage.sol:9:17:\n  |\n9 | contrat Storage {\n  |                 ^\n\n",
+      );
     chai.expect(jobRes.body.contract).to.deep.equal({
       match: null,
       creationMatch: null,
@@ -301,61 +214,106 @@ describe("POST /v2/verify/:chainId/:address", function () {
     });
   });
 
-  it("should return a 429 if the contract is being verified at the moment already", async () => {
-    makeWorkersWait();
+  it("should store a job error if the Vyper compiler returns an error", async () => {
+    const { resolveWorkers } = makeWorkersWait();
 
-    // First verification request
-    const verifyRes1 = await chai
+    const vyperContractPath = path.join(
+      __dirname,
+      "..",
+      "..",
+      "..",
+      "sources",
+      "vyper",
+      "testcontract3_fail",
+    );
+    const vyperArtifactPath = path.join(vyperContractPath, "artifact.json");
+    const vyperArtifact = JSON.parse(
+      fs.readFileSync(vyperArtifactPath, "utf8"),
+    );
+    const vyperSourceFileName = "test.vy";
+    const vyperSourcePath = path.join(vyperContractPath, vyperSourceFileName);
+    const vyperSource = fs.readFileSync(vyperSourcePath, "utf8");
+
+    const { contractAddress, txHash } =
+      await deployFromAbiAndBytecodeForCreatorTxHash(
+        chainFixture.localSigner,
+        vyperArtifact.abi,
+        vyperArtifact.bytecode,
+      );
+
+    const verifyRes = await chai
       .request(serverFixture.server.app)
-      .post(
-        `/v2/verify/${chainFixture.chainId}/${chainFixture.defaultContractAddress}`,
-      )
+      .post(`/v2/verify/${chainFixture.chainId}/${contractAddress}`)
       .send({
-        stdJsonInput: chainFixture.defaultContractJsonInput,
-        compilerVersion:
-          chainFixture.defaultContractMetadataObject.compiler.version,
-        contractIdentifier: Object.entries(
-          chainFixture.defaultContractMetadataObject.settings.compilationTarget,
-        )[0].join(":"),
-        creationTransactionHash: chainFixture.defaultContractCreatorTx,
+        stdJsonInput: {
+          language: "Vyper",
+          sources: {
+            [vyperSourceFileName]: {
+              content: vyperSource,
+            },
+          },
+          settings: {
+            evmVersion: "istanbul",
+            outputSelection: {
+              "*": ["evm.bytecode"],
+            },
+          },
+        },
+        compilerVersion: "0.3.10+commit.91361694",
+        contractIdentifier: `${vyperSourceFileName}:${vyperSourceFileName.split(".")[0]}`,
+        creationTransactionHash: txHash,
       });
+    chai.expect(verifyRes.status).to.equal(202);
 
-    chai.expect(verifyRes1.status).to.equal(202);
+    await resolveWorkers();
 
-    // Second verification request before the first one completes
-    const verifyRes2 = await chai
+    const jobRes = await chai
       .request(serverFixture.server.app)
-      .post(
-        `/v2/verify/${chainFixture.chainId}/${chainFixture.defaultContractAddress}`,
-      )
-      .send({
-        stdJsonInput: chainFixture.defaultContractJsonInput,
-        compilerVersion:
-          chainFixture.defaultContractMetadataObject.compiler.version,
-        contractIdentifier: Object.entries(
-          chainFixture.defaultContractMetadataObject.settings.compilationTarget,
-        )[0].join(":"),
-        creationTransactionHash: chainFixture.defaultContractCreatorTx,
-      });
+      .get(`/v2/verify/${verifyRes.body.verificationId}`);
 
-    chai.expect(verifyRes2.status).to.equal(429);
+    chai.expect(jobRes.status).to.equal(200);
+    chai.expect(jobRes.body).to.include({
+      isJobCompleted: true,
+    });
+    chai.expect(jobRes.body.error).to.exist;
+    chai.expect(jobRes.body.error.customCode).to.equal("compiler_error");
     chai
-      .expect(verifyRes2.body.customCode)
-      .to.equal("duplicate_verification_request");
-    chai.expect(verifyRes2.body).to.have.property("errorId");
-    chai.expect(verifyRes2.body).to.have.property("message");
+      .expect(jobRes.body.error.errorData.compilerErrors[0].formattedMessage)
+      .to.equal(
+        'invalid syntax (<unknown>, line 11)\n  line 11:22 \n       10 def helloWorld() -> String[24]:\n  ---> 11     error_in_the_code!\n  ------------------------------^\n       12     return "Hello Vyper!"\n',
+      );
+    chai.expect(jobRes.body.contract).to.deep.equal({
+      match: null,
+      creationMatch: null,
+      runtimeMatch: null,
+      chainId: chainFixture.chainId,
+      address: contractAddress,
+    });
+  });
+
+  it("should return a 429 if the contract is being verified at the moment already", async () => {
+    await testAlreadyBeingVerified(
+      serverFixture,
+      makeWorkersWait,
+      `/v2/verify/${chainFixture.chainId}/${chainFixture.defaultContractAddress}`,
+      {
+        stdJsonInput: chainFixture.defaultContractJsonInput,
+        compilerVersion:
+          chainFixture.defaultContractMetadataObject.compiler.version,
+        contractIdentifier: Object.entries(
+          chainFixture.defaultContractMetadataObject.settings.compilationTarget,
+        )[0].join(":"),
+        creationTransactionHash: chainFixture.defaultContractCreatorTx,
+      },
+    );
   });
 
   it("should return a 409 if the contract is already verified", async () => {
-    const { resolveWorkers, runTaskStub } = makeWorkersWait();
-
-    // First verification request
-    const verifyRes1 = await chai
-      .request(serverFixture.server.app)
-      .post(
-        `/v2/verify/${chainFixture.chainId}/${chainFixture.defaultContractAddress}`,
-      )
-      .send({
+    await testAlreadyVerified(
+      serverFixture,
+      makeWorkersWait,
+      `/v2/verify/${chainFixture.chainId}/${chainFixture.defaultContractAddress}`,
+      {
         stdJsonInput: chainFixture.defaultContractJsonInput,
         compilerVersion:
           chainFixture.defaultContractMetadataObject.compiler.version,
@@ -363,51 +321,10 @@ describe("POST /v2/verify/:chainId/:address", function () {
           chainFixture.defaultContractMetadataObject.settings.compilationTarget,
         )[0].join(":"),
         creationTransactionHash: chainFixture.defaultContractCreatorTx,
-      });
-
-    chai.expect(verifyRes1.status).to.equal(202);
-
-    await resolveWorkers();
-    const contractRes = await chai
-      .request(serverFixture.server.app)
-      .get(
-        `/v2/contract/${chainFixture.chainId}/${chainFixture.defaultContractAddress}`,
-      );
-
-    chai.expect(contractRes.status).to.equal(200);
-    chai.expect(contractRes.body).to.include({
-      match: "exact_match",
-      creationMatch: "exact_match",
-      runtimeMatch: "exact_match",
-      chainId: chainFixture.chainId,
-      address: chainFixture.defaultContractAddress,
-    });
-
-    runTaskStub.restore();
-    const { resolveWorkers: resolveWorkers2 } = makeWorkersWait();
-
-    // Second verification
-    const verifyRes2 = await chai
-      .request(serverFixture.server.app)
-      .post(
-        `/v2/verify/${chainFixture.chainId}/${chainFixture.defaultContractAddress}`,
-      )
-      .send({
-        stdJsonInput: chainFixture.defaultContractJsonInput,
-        compilerVersion:
-          chainFixture.defaultContractMetadataObject.compiler.version,
-        contractIdentifier: Object.entries(
-          chainFixture.defaultContractMetadataObject.settings.compilationTarget,
-        )[0].join(":"),
-        creationTransactionHash: chainFixture.defaultContractCreatorTx,
-      });
-
-    chai.expect(verifyRes2.status).to.equal(409);
-    chai.expect(verifyRes2.body.customCode).to.equal("already_verified");
-    chai.expect(verifyRes2.body).to.have.property("errorId");
-    chai.expect(verifyRes2.body).to.have.property("message");
-
-    await resolveWorkers2();
+      },
+      chainFixture.chainId,
+      chainFixture.defaultContractAddress,
+    );
   });
 
   it("should return a 400 if the standard json input misses the language", async () => {
@@ -552,5 +469,78 @@ describe("POST /v2/verify/:chainId/:address", function () {
     chai.expect(verifyRes.body.customCode).to.equal("invalid_parameter");
     chai.expect(verifyRes.body).to.have.property("errorId");
     chai.expect(verifyRes.body).to.have.property("message");
+  });
+
+  it("should return a 404 when the chain is not found", async function () {
+    const unknownChainId = "5";
+    const chainMap = serverFixture.server.chainRepository.sourcifyChainMap;
+    sandbox.stub(chainMap, unknownChainId).value(undefined);
+
+    const verifyRes = await chai
+      .request(serverFixture.server.app)
+      .post(
+        `/v2/verify/${unknownChainId}/${chainFixture.defaultContractAddress}`,
+      )
+      .send({
+        stdJsonInput: chainFixture.defaultContractJsonInput,
+        compilerVersion:
+          chainFixture.defaultContractMetadataObject.compiler.version,
+        contractIdentifier: Object.entries(
+          chainFixture.defaultContractMetadataObject.settings.compilationTarget,
+        )[0].join(":"),
+        creationTransactionHash: chainFixture.defaultContractCreatorTx,
+      });
+
+    chai.expect(verifyRes.status).to.equal(404);
+    chai.expect(verifyRes.body.customCode).to.equal("unsupported_chain");
+    chai.expect(verifyRes.body).to.have.property("errorId");
+    chai.expect(verifyRes.body).to.have.property("message");
+  });
+
+  it("should fail matching with creation tx if the provided creationTransactionHash does not match the contract address", async () => {
+    // Deploy contract A
+    const deploymentA = await deployFromAbiAndBytecodeForCreatorTxHash(
+      chainFixture.localSigner,
+      chainFixture.defaultContractArtifact.abi,
+      chainFixture.defaultContractArtifact.bytecode,
+    );
+
+    // Deploy contract B
+    const deploymentB = await deployFromAbiAndBytecodeForCreatorTxHash(
+      chainFixture.localSigner,
+      chainFixture.defaultContractArtifact.abi,
+      chainFixture.defaultContractArtifact.bytecode,
+    );
+
+    const { resolveWorkers } = makeWorkersWait();
+
+    // Try to verify contract A, but provide B's creatorTxHash
+    const verifyRes = await chai
+      .request(serverFixture.server.app)
+      .post(`/v2/verify/${chainFixture.chainId}/${deploymentA.contractAddress}`)
+      .send({
+        stdJsonInput: chainFixture.defaultContractJsonInput,
+        compilerVersion:
+          chainFixture.defaultContractMetadataObject.compiler.version,
+        contractIdentifier: Object.entries(
+          chainFixture.defaultContractMetadataObject.settings.compilationTarget,
+        )[0].join(":"),
+        creationTransactionHash: deploymentB.txHash,
+      });
+
+    await resolveWorkers();
+
+    // Fetch the job result
+    const jobRes = await chai
+      .request(serverFixture.server.app)
+      .get(`/v2/verify/${verifyRes.body.verificationId}`);
+
+    chai.expect(jobRes.status).to.be.oneOf([200]);
+    chai.expect(jobRes.body).to.include({
+      isJobCompleted: true,
+    });
+    chai.expect(jobRes.body.error).to.not.exist;
+    chai.expect(jobRes.body.contract.creationMatch).to.be.null;
+    chai.expect(jobRes.body.contract.runtimeMatch).to.equal("exact_match");
   });
 });